core: Lower the `WorkingCopy` trait
This is in preparation to move `Workspace` over to the new core crate.

TODO: this is incomplete and not building since `SnapshotOptions` includes a `GitIgnore` which isn't ported yet.
(this probably should just be called `Ignore` if its in core, see my FR for a `.jj-ignore`)
