+++ core/src/ignore.rs
